+++ core/server/common/src/main/java/alluxio/master/journal/ufs/UfsJournal.java
@@ -698,20 +698,5 @@ public class UfsJournal implements Journal {
       // Update suspended sequence after catch-up is finished.
       mSuspendSequence = catchUp(mCatchUpStartSequence, mCatchUpEndSequence) - 1;
     }
-
-    @Override
-    public void onError(Throwable t) {
-      if (ExceptionUtils.containsInterruptedException(t)) {
-        // Tolerate interruption when the master is failing over or closing
-        // so we don't extra-crash
-        LOG.warn("Thread {} interrupted, assume the master is failing over or shutting down",
-            Thread.currentThread().getId());
-        return;
-      }
-      LOG.error("Uncaught exception from thread {}", Thread.currentThread().getId(), t);
-      setError(t);
-      // If the journal catchup fails, crash the standby master
-      ProcessUtils.fatalError(LOG, t, "Failed to catch up journal");
-    }
   }
 }
